Utils/xlogs/pyproject.toml
[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core"]

[tool.poetry]
authors = ["Your Name <you@example.com>"]
description = ""
name = "xlogs"
readme = "README.md"
version = "0.1.0"

[tool.poetry.dependencies]
google-cloud-storage = "^2.18.2"
loguru = "^0.7.2"
more-itertools = "^10.5.0"
python = "^3.10"
tabulate = "^0.9.0"
typer = "^0.15.1"

[tool.poetry.scripts]
xlogs = "xlogs.main:app"
